Refactor - Use less `#[cfg()]` magic
- Split platform dependent signal implementations into
  `mod unix` and `mod windows`.

- Use `pending` for those signals that are not supported on Windows.
